+++ CHANGELOG.md
@@ -14,12 +14,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-* Improved content negotiation and AUTHORIZED_FETCH support for third-party plugins
+* Improved content negotiation and AUTHORIZED_FETCH support for third-party plugins.
+* Moved password check to `is_post_disabled` function.
 
 ### Fixed
 
 * Handle deletes from remote servers that leave behind an accessible Tombstone object.
 * No longer parses tags for post types that don't support Activitypub.
+* rel attribute will now contain no more than one "me" value.
 
 ## [4.7.3] - 2025-01-21